fix(tx_builder): `add_utxo` requires explicit opt-in to select spent outputs
Having `params.bumping_fee` mediate replaceability was too broad
in that it would allow selecting any previously spent output
regardless of whether we opted into replacing the original tx.
Instead we introduce `TxParams::replacing` member to keep track of
the outpoints that may be re-selected for an RBF.

The implementation of `replace_tx` is changed to populate the
`replacing` set with all of the original outpoints. This signals
to the TxBuilder that any of the original inputs may be replaced,
although we still only pick one as the argument to `add_utxo`.